Fixed warnings

+++ src/graphics/VertexArray.cpp
@@ -1,5 +1,5 @@
-#include "GL/glew.h"
 #include "graphics/VertexArray.hpp"
+#include "GL/glew.h"
 
 namespace renderel::graphics {
 
@@ -23,9 +23,10 @@ void VertexArray::AddBuffer(const VertexBuffer &vb,
 	for (unsigned int i = 0; i < elements.size(); ++i) {
 		const auto &element = elements[i];
 		GLCall(glEnableVertexAttribArray(i));
-		GLCall(glVertexAttribPointer(i, element.count, element.type,
-									 element.normalized, layout.GetStride(),
-									 (const void *)offset));
+        GLCall(glVertexAttribPointer(i, static_cast<int>(element.count),
+                                     element.type, element.normalized,
+                                     static_cast<int>(layout.GetStride()),
+                                     (const void *)offset));
 		offset +=
 			element.count * VertexBufferElement::GetSizeOfType(element.type);
 	}
